# contracts/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: DSH 工业仿真智能体 工程 API
  version: 0.1.2
  description: |
    dsh-sim R0 工程能力真相入口（定义书 §工程API：业务操作 / §工程API：审查、节点与上传）。
    MCP 是本 API 的受限投影。长操作返回 202 与资源 ID；请求返回不是计算结束。
    所有产生副作用的 POST 必须携带 Idempotency-Key 头（主体+项目+action+key 唯一；
    同 key 同摘要返回原对象，同 key 异摘要 → 409）。
    错误对象统一为 Error{code,message,retryable,trace_id,details}。
    本文件中的示例值均为占位：假哈希一律写作 "sha256:<64hex>" 形式，不含真实工程参数。
servers:
  - url: http://127.0.0.1:8600/api/v1
...
paths:
  /capabilities:
    get:
      tags: [capabilities]
      operationId: listCapabilities
      summary: 查询当前域已发布能力包目录（游标分页）
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 能力包列表页
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CapabilityPackage' }
                  next_cursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /tasks/{task_id}/bundles:
    post:
      tags: [bundles]
      operationId: buildBundle
      summary: 异步构建冻结证据集合；检查必需 Run 与定义一致性
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision: { type: integer, minimum: 1 }
      responses:
        '201':
          description: 冻结证据包已同步构建（缺必需工况标 incomplete 并阻塞 ACCEPT）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bundle' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /bundles/{bundle_id}:
    get:
      tags: [bundles]
      operationId: getBundle
      summary: 读取冻结证据包清单与摘要
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Bundle 对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bundle' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/reviews:
    post:
      tags: [reviews]
      operationId: submitReview
      summary: 提交冻结包给预先指定的审查人；提交动作不等于接受
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision, bundle_id, bundle_digest, reviewer_id]
              properties:
                revision: { type: integer, minimum: 1 }
                bundle_id: { type: string }
                bundle_digest:
                  type: string
                  example: "sha256:<64hex>"
                reviewer_id: { type: string }
      responses:
        '201':
          description: 审查已创建（PENDING）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /reviews/{review_id}:
    get:
      tags: [reviews]
      operationId: getReview
      summary: 返回审查问题与有效性
      parameters:
        - name: review_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Review 对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /reviews/{review_id}/issues:
    post:
      tags: [issues]
      operationId: createIssue
      summary: 创建问题；模型创建的为 DRAFT，经 confirmIssue 人工确认转 OPEN
      parameters:
        - name: review_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [responsible, severity, description, close_criteria]
              properties:
                responsible: { type: string }
                severity: { type: string, example: "<severity>" }
                description: { type: string }
                close_criteria: { type: string }
                related_artifact_ids:
                  type: array
                  items: { type: string }
                related_run_ids:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: 问题已创建（DRAFT）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewIssue' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /issues/{issue_id}/replies:
    post:
      tags: [issues]
      operationId: replyIssue
      summary: 回复问题并附新增证据；答复文本本身不能自动关闭问题
      parameters:
        - name: issue_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                evidence_artifact_ids:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: 回复已记录
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewIssue' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /issues/{issue_id}/confirm:
    post:
      tags: [issues]
      operationId: confirmIssue
      summary: 指定审查人人工确认 DRAFT→OPEN；模型接口无此工具
      parameters:
        - name: issue_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issue_version, confirmation_id]
              properties:
                issue_version: { type: integer, minimum: 1 }
                confirmation_id: { type: string }
      responses:
        '200':
          description: 问题已转 OPEN
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewIssue' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /issues/{issue_id}/close:
    post:
      tags: [issues]
      operationId: closeIssue
      summary: 审查人核对回复与新增证据后关闭；一次性人工确认
      parameters:
        - name: issue_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issue_version, close_evidence_artifact_ids, confirmation_id]
              properties:
                issue_version: { type: integer, minimum: 1 }
                close_evidence_artifact_ids:
                  type: array
                  items: { type: string }
                confirmation_id: { type: string }
      responses:
        '200':
          description: 问题已关闭
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewIssue' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /reviews/{review_id}/decisions:
    post:
      tags: [reviews]
      operationId: decideReview
      summary: 人工决定；服务端必检包摘要、修订有效性、Run 完整性、各项检查、无未关闭阻塞、职责分离
      parameters:
        - name: review_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome, bundle_digest, revision, confirmation_id]
              properties:
                outcome:
                  enum: [ACCEPT, REQUEST_CHANGES, REJECT]
                bundle_digest:
                  type: string
                  example: "sha256:<64hex>"
                revision: { type: integer, minimum: 1 }
                limitations: { type: string }
                confirmation_id: { type: string }
      responses:
        '201':
          description: 决定已记录（不可变）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Decision' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /confirmations:
    post:
      tags: [confirmations]
      operationId: createHumanConfirmation
      summary: 受信 UI 会话发起一次性人工确认；绑定 action/target/digest，短期有效、单次消费，不向模型泄露凭据
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, target_id, target_digest]
              properties:
                action: { type: string, example: "<action e.g. authorizeRuns>" }
                target_id: { type: string }
                target_digest:
                  type: string
                  example: "sha256:<64hex>"
      responses:
        '201':
          description: 确认凭据已签发（短期有效、单次消费）
          content:
            application/json:
              schema:
                type: object
                required: [confirmation_id, expires_at]
                properties:
                  confirmation_id: { type: string }
                  expires_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: 幂等键；主体+项目+action+key 唯一。同 key 同请求摘要返回原对象；同 key 异摘要 → 409。
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: 游标分页：上一页返回的 next_cursor。
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: 401 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 403 越权
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 409 摘要/修订/幂等冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: 422 工程输入不成立
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: 429 配额限制
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServiceUnavailable:
      description: 503 暂不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, trace_id, details]
      properties:
        code:
          type: string
          description: 字符串枚举错误码
          example: "<ERROR_CODE>"
        message: { type: string }
        retryable:
          type: boolean
          description: 客户端只能对 retryable 与已定义状态采取有界重试
        trace_id: { type: string, format: uuid }
        details: { type: object }
    Sha256:
      type: string
      description: SHA-256 摘要（64 位小写十六进制）；占位写作 sha256:<64hex>。哈希仅由服务计算。
      example: "sha256:<64hex>"
...
    ReviewState:
      enum: [NOT_SUBMITTED, PENDING, CHANGES_REQUESTED, ACCEPTED, REJECTED]
    ValidityState:
      enum: [CURRENT, STALE]
...
    Bundle:
      type: object
      required: [bundle_id, task_id, revision, bundle_digest, manifest, validity, created_at]
      properties:
        bundle_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        bundle_digest: { $ref: '#/components/schemas/Sha256' }
        manifest:
          type: array
          items:
            type: object
            required: [artifact_id, logical_path, sha256]
            properties:
              artifact_id: { type: string }
              logical_path: { type: string }
              sha256: { $ref: '#/components/schemas/Sha256' }
          description: 冻结清单，不可原位改写
        validity: { $ref: '#/components/schemas/ValidityState' }
        created_at: { type: string, format: date-time }
...
    Review:
      type: object
      required: [review_id, task_id, revision, bundle_id, bundle_digest, reviewer_id, state, validity, issues, created_at]
      properties:
        review_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        bundle_id: { type: string }
        bundle_digest: { $ref: '#/components/schemas/Sha256' }
        reviewer_id: { type: string }
        state: { $ref: '#/components/schemas/ReviewState' }
        validity: { $ref: '#/components/schemas/ValidityState' }
        issues:
          type: array
          items: { $ref: '#/components/schemas/ReviewIssue' }
        created_at: { type: string, format: date-time }
    ReviewIssue:
      type: object
      required: [issue_id, review_id, created_by, responsible, severity, description, close_criteria, status, version, created_at]
      properties:
        issue_id: { type: string }
        review_id: { type: string }
        created_by: { type: string }
        responsible: { type: string }
        severity: { type: string }
        description: { type: string }
        close_criteria: { type: string }
        related_artifact_ids:
          type: array
          items: { type: string }
        related_run_ids:
          type: array
          items: { type: string }
        replies:
          type: array
          items:
            type: object
            required: [reply_id, author_id, body, created_at]
            properties:
              reply_id: { type: string }
              author_id: { type: string }
              body: { type: string }
              evidence_artifact_ids:
                type: array
                items: { type: string }
              created_at: { type: string, format: date-time }
        status:
          enum: [DRAFT, OPEN, CLOSED]
          description: 模型创建的为 DRAFT；confirmIssue 人工确认转 OPEN；closeIssue 需审查人确认
        version: { type: integer, minimum: 1 }
        closed_by: { type: string, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    Decision:
      type: object
      required: [decision_id, review_id, outcome, decided_by, task_id, revision, bundle_digest, purpose, decided_at]
      properties:
        decision_id: { type: string }
        review_id: { type: string }
        outcome:
          enum: [ACCEPT, REQUEST_CHANGES, REJECT]
          description: ACCEPT 在数值 FAIL/证据 INSUFFICIENT/范围 UNCONFIRMED/MOCK/未关闭阻塞时必须失败
        decided_by: { type: string, description: 指定审查人；职责分离，不能自行验收 }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        bundle_digest: { $ref: '#/components/schemas/Sha256' }
        purpose: { type: string }
        limitations: { type: string, nullable: true }
        decided_at: { type: string, format: date-time }
      description: 已接受决定保持不可变；新输入使其对当前输入失效而非删除。
    CapabilityPackage:
      type: object
      required: [capability_package_id, version, manifest_sha256, status, purpose]
      properties:
        capability_package_id: { type: string }
        version: { type: string }
        manifest_sha256: { $ref: '#/components/schemas/Sha256' }
        status:
          enum: [DRAFT, RELEASED, RETIRED]
          description: 只有 RELEASED 且与当前软件匹配的能力可用于正式任务
        purpose: { type: string }
        domain_summary: { type: string, nullable: true }
        compatibility:
          type: object
          nullable: true
          description: 软件版本兼容信息（占位）